+++ NEWS.md
@@ -17,6 +17,7 @@
 * Added support of the full screen for a `module` when the `filters` argument is equal `NULL`.
 * Added support for `shiny::moduleServer` passed to the `server` parameter of `teal::module`.
 * Added `pid` into footnote so app developers can identify logs for apps.
+* Added logs to the modules and sub-modules returned from `init`.
 
 ### Bug fixes
 * Fixed data loading of `DatasetConnector` being dependent on other `Dataset` or `DatasetConnector` objects.